Introduce the AuthStoring type.
This protocol makes it possible to intercept authentication storage and retrieval on behalf of the PorscheConnect class. This can be used to store and retrieve authentication tokens from a cache (typically disk), potentially reducing the number of needed authentication API calls across multiple sessions.

By default, PorscheConnect uses a simple in-memory auth storage mechanism that will lose auth tokens when the PorscheConnect instance is deallocated.