Allow auth token storage to be async.
This will allow the auth storage to be synchronized using the `actor` type.

This change required significant syntactic updates to the tests. Unfortunately, XCTest does not yet support `async`, so a third party library was added to the test target as a shim. If/when Apple adds support for async XCTAssert statements, it should be a simple find-replace operation to update to the modern APIs.